# docs/api/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "AURA Care-Chain API",
        "description": "API REST do AURA Care-Chain \u2014 assistente de sa\u00fade domiciliar voice-first\npara idosos com Parkinson, com cadeia log\u00edstica de seguran\u00e7a da casa.\n\nConsumida por tr\u00eas clientes: app Flutter (Android/iOS/Web),\napp React Native e painel administrativo Angular.\n\nRegras de ouro: nunca prescreve nem diagnostica, o escore \u00e9 sempre\nexplic\u00e1vel (fatores + pesos) e nenhum dado de sa\u00fade \u00e9 gravado sem\nconsentimento LGPD.\n",
        "contact": {
            "name": "Equipe AURA \u2014 FIAP Smart HAS"
        },
        "license": {
            "name": "Uso acad\u00eamico \u2014 FIAP 2026"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/v1/recommendations": {
            "post": {
                "tags": [
                    "5. Care-Chain"
                ],
                "summary": "Gera a recomenda\u00e7\u00e3o explicada a partir do escore da casa",
                "operationId": "recommend",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRecommendationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecommendationResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha inesperada \u2014 a mensagem interna fica no log do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "INTERNAL_ERROR": {
                                        "value": {
                                            "error": {
                                                "code": "INTERNAL_ERROR",
                                                "message": "Erro interno.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente, inv\u00e1lido ou expirado \u2014 ou credenciais incorretas no login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "UNAUTHORIZED": {
                                        "value": {
                                            "error": {
                                                "code": "UNAUTHORIZED",
                                                "message": "Autentica\u00e7\u00e3o necess\u00e1ria.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "NOT_FOUND": {
                                        "value": {
                                            "error": {
                                                "code": "NOT_FOUND",
                                                "message": "Casa n\u00e3o encontrada(a).",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Papel sem permiss\u00e3o (RBAC) ou recurso de outro paciente (isolamento)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "FORBIDDEN": {
                                        "value": {
                                            "error": {
                                                "code": "FORBIDDEN",
                                                "message": "Acesso negado a este recurso.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Escore n\u00e3o pertence \u00e0 casa, sem produto para o risco, ou consentimento ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "NO_PRODUCT": {
                                        "value": {
                                            "error": {
                                                "code": "NO_PRODUCT",
                                                "message": "Sem produto no cat\u00e1logo para o risco 'fall_bathroom'.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/recommendations/{recommendationId}/reject": {
            "post": {
                "tags": [
                    "5. Care-Chain"
                ],
                "summary": "Recusa a recomenda\u00e7\u00e3o \u2014 nenhum pedido \u00e9 criado",
                "operationId": "reject",
                "parameters": [
                    {
                        "name": "recommendationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecommendationResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha inesperada \u2014 a mensagem interna fica no log do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "INTERNAL_ERROR": {
                                        "value": {
                                            "error": {
                                                "code": "INTERNAL_ERROR",
                                                "message": "Erro interno.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente, inv\u00e1lido ou expirado \u2014 ou credenciais incorretas no login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "UNAUTHORIZED": {
                                        "value": {
                                            "error": {
                                                "code": "UNAUTHORIZED",
                                                "message": "Autentica\u00e7\u00e3o necess\u00e1ria.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "NOT_FOUND": {
                                        "value": {
                                            "error": {
                                                "code": "NOT_FOUND",
                                                "message": "Casa n\u00e3o encontrada(a).",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Papel sem permiss\u00e3o (RBAC) ou recurso de outro paciente (isolamento)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "FORBIDDEN": {
                                        "value": {
                                            "error": {
                                                "code": "FORBIDDEN",
                                                "message": "Acesso negado a este recurso.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/recommendations/{recommendationId}/approve": {
            "post": {
                "tags": [
                    "5. Care-Chain"
                ],
                "summary": "Aprova\u00e7\u00e3o humana da cuidadora \u2014 \u00fanica porta que cria o pedido",
                "operationId": "approve",
                "parameters": [
                    {
                        "name": "recommendationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApproveResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha inesperada \u2014 a mensagem interna fica no log do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "INTERNAL_ERROR": {
                                        "value": {
                                            "error": {
                                                "code": "INTERNAL_ERROR",
                                                "message": "Erro interno.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente, inv\u00e1lido ou expirado \u2014 ou credenciais incorretas no login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "UNAUTHORIZED": {
                                        "value": {
                                            "error": {
                                                "code": "UNAUTHORIZED",
                                                "message": "Autentica\u00e7\u00e3o necess\u00e1ria.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "NOT_FOUND": {
                                        "value": {
                                            "error": {
                                                "code": "NOT_FOUND",
                                                "message": "Casa n\u00e3o encontrada(a).",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Papel sem permiss\u00e3o (RBAC) ou recurso de outro paciente (isolamento)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "FORBIDDEN": {
                                        "value": {
                                            "error": {
                                                "code": "FORBIDDEN",
                                                "message": "Acesso negado a este recurso.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Recomenda\u00e7\u00e3o j\u00e1 aprovada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "CONFLICT": {
                                        "value": {
                                            "error": {
                                                "code": "CONFLICT",
                                                "message": "Recomenda\u00e7\u00e3o j\u00e1 aprovada.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Recomenda\u00e7\u00e3o rejeitada n\u00e3o vira pedido (RN-022)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "APPROVAL_REQUIRED": {
                                        "value": {
                                            "error": {
                                                "code": "APPROVAL_REQUIRED",
                                                "message": "Recomenda\u00e7\u00e3o rejeitada n\u00e3o vira pedido.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{orderId}/advance": {
            "post": {
                "tags": [
                    "5. Care-Chain"
                ],
                "summary": "Avan\u00e7a o pedido: approved \u2192 sourcing \u2192 in_route \u2192 delivered \u2192 installed \u2192 returned",
                "operationId": "advance",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdvanceResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha inesperada \u2014 a mensagem interna fica no log do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "INTERNAL_ERROR": {
                                        "value": {
                                            "error": {
                                                "code": "INTERNAL_ERROR",
                                                "message": "Erro interno.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente, inv\u00e1lido ou expirado \u2014 ou credenciais incorretas no login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "UNAUTHORIZED": {
                                        "value": {
                                            "error": {
                                                "code": "UNAUTHORIZED",
                                                "message": "Autentica\u00e7\u00e3o necess\u00e1ria.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "NOT_FOUND": {
                                        "value": {
                                            "error": {
                                                "code": "NOT_FOUND",
                                                "message": "Casa n\u00e3o encontrada(a).",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Papel sem permiss\u00e3o (RBAC) ou recurso de outro paciente (isolamento)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "FORBIDDEN": {
                                        "value": {
                                            "error": {
                                                "code": "FORBIDDEN",
                                                "message": "Acesso negado a este recurso.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Pedido em est\u00e1gio terminal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "CONFLICT": {
                                        "value": {
                                            "error": {
                                                "code": "CONFLICT",
                                                "message": "Pedido em est\u00e1gio terminal.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateRecommendationRequest": {
                "required": [
                    "homeId"
                ],
                "type": "object",
                "properties": {
                    "homeId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "scoreId": {
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "RecommendationResponse": {
                "type": "object",
                "properties": {
                    "recommendationId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "sku": {
                        "type": "string"
                    },
                    "productName": {
                        "type": "string"
                    },
                    "reason": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "factors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "weights": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                },
                "description": "Recomenda\u00e7\u00e3o explicada. `reason` \u00e9 o texto que a cuidadora l\u00ea antes de aprovar e passa\npelo guardrail de n\u00e3o-prescri\u00e7\u00e3o. Enquanto `status` for `recommended`, nenhum pedido existe:\ns\u00f3 a aprova\u00e7\u00e3o humana cria o pedido (RN-022)."
            },
            "ApproveResponse": {
                "type": "object",
                "properties": {
                    "orderId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "stage": {
                        "type": "string",
                        "enum": [
                            "approved",
                            "sourcing",
                            "in_route",
                            "delivered",
                            "installed",
                            "returned"
                        ]
                    }
                }
            },
            "AdvanceResponse": {
                "type": "object",
                "properties": {
                    "stage": {
                        "type": "string",
                        "enum": [
                            "approved",
                            "sourcing",
                            "in_route",
                            "delivered",
                            "installed",
                            "returned"
                        ]
                    },
                    "etaDelivery": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "installAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "slaBreached": {
                        "type": "boolean"
                    }
                }
            },
...
            "ApiErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/Body"
                    }
                }
            },
            "Body": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "details": {
                        "type": "object"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido em POST /api/v1/auth/login",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}